Fixed return code got mistakenly cast from double to int error in spectrum.hpp. https://tracker.physiomeproject.org/show_bug.cgi?id=3636

+++ core/source/api/zinc/spectrum.hpp
@@ -161,7 +161,7 @@ public:
 		return Cmiss_spectrum_component_set_colour_maximum(id, value);
 	}
 
-	int getStepValue()
+	double getStepValue()
 	{
 		return Cmiss_spectrum_component_get_step_value(id);
 	}
@@ -171,7 +171,7 @@ public:
 		return Cmiss_spectrum_component_set_step_value(id, value);
 	}
 
-	int getExaggeration()
+	double getExaggeration()
 	{
 		return Cmiss_spectrum_component_get_exaggeration(id);
 	}
@@ -181,7 +181,7 @@ public:
 		return Cmiss_spectrum_component_set_exaggeration(id, value);
 	}
 
-	int getBandedRatio()
+	double getBandedRatio()
 	{
 		return Cmiss_spectrum_component_get_banded_ratio(id);
 	}
